(#108, #113) Moved images/models to resources package
