+++ src/example/java/tech/fastj/example/modelreadwrite/Main.java
@@ -5,7 +5,7 @@ import tech.fastj.math.Pointf;
 import tech.fastj.graphics.game.Model2D;
 import tech.fastj.graphics.game.Polygon2D;
 import tech.fastj.graphics.util.DrawUtil;
-import tech.fastj.graphics.util.ModelUtil;
+import tech.fastj.resources.models.ModelUtil;
 
 import java.io.IOException;
 import java.nio.file.Files;
